requirements.txt
langchain
requests
beautifulsoup4
serpapi
python-dotenv
textblob
google-search-results
